LayoutTests/accessibility/win/list-marker-role-expected.txt

This test verifies that, on Windows, list markers return the non-editable "text" role.


list item
PASS listMarker.role is "AXRole: AXStaticText"
